[BUGFIX] Do not fetch extension list on every request
Due to comparing a DateTime object with an integer
the extension list is fetched every time when switching
to the get extensions tab in the extension manager.

Convert the DateTime object to a timestamp first and
add a test which ensure this works.

Additionally clean up the doc comment of that
property in the model and use forceUpdateCheck
as action parameter in the controller.

Change-Id: I71f6599a6fa41115515c1764deb26e8e56762f4f
Resolves: #39944
Releases: 6.0
Reviewed-on: http://review.typo3.org/13905
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog