[BUGFIX] EM does not update extension list
Currently the Extension Manager looks up the field
in the extension repository repository on how many
extensions are in the system.

Since the result of this getter might be a string
or the number in there might not be up to date,
the EM does not update the extension list.

Change that to a count on the extension table,
which is the wanted behavior here.

Fixes: #43460
Releases: 6.0, 6.1

Change-Id: I3917148852995d0853fbca60848881d6aeacc989
Reviewed-on: http://review.typo3.org/16857
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn