[BUGFIX] EM ConstantsView crashs with dep/faulty types.
The types default and input are deprecated. Removing of this let
the ConstantsView crash, as there is no function configuration for them.

We now readd the config in 6.1 as we have no problems with this lines
in the core and we don't know how much extsnions exists which fail.

Resolves: #46996
Related: #44763
Releases: 6.1, 6.0
Change-Id: I670e80bab6b415997fcfd26de19fba870a0f8446
Reviewed-on: https://review.typo3.org/19697
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn