[BUGFIX] Removed extension stays in PackageStates file
Deleting an extension in the Extension Manager doesn't delete
the entry in PackageStates file.

Use the package manager to delete the extension. This will take care of
linked extensions (just remove the link) too.
The package manager will take care to remove the entry from the
PackageStates file.

If an extension is not listed in the PackageStates file (e.g. because
we uploaded it via FTP and forgot to clear the system caches),
directly remove the folder.

Resolves: #53459
Releases: 6.2
Change-Id: I10d5ef33ccf6438fece512478408c57d5d7486de
Reviewed-on: https://review.typo3.org/26921
Reviewed-by: Jan Runte
Tested-by: Jan Runte
Reviewed-by: Helmut Hummel
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn