[TASK] EM: Add possibility to bypass system dependency checks
This patch extends the extension manager in three different ways. First
all dependencies are checked and error messages are bundled to show all
problems to the user.

Secondly on uploading an extension file the installation process is
started automatically.

The main change of this patch is to introduce a new function to prevent
(system) dependency checks. This means the checks for TYPO3 and PHP
version don't throw an exception anymore. Required extensions are tried
to be fetched from TER but don't stop installation either.

If errors occur on first installation process, a link to force the
installation is added to the notification. Before any installation
can be run, a dialog with a "break warning" is shown and has to be
confirmed by clicking the unfocussed field.

To be able to skip the system dependency check, a new property for
ExtensionManagementService and DependencyUtility is introduced which
controls disabling the system dependency check. All extension
dependencies are still resolved and needed extensions are tried
to be fetched from TER.

Resolves: #54512
Releases: 6.2
Change-Id: Ia11b7770a2773538bda48d889282ff51bf187c84
Reviewed-on: https://review.typo3.org/28924
Reviewed-by: Sascha Wilking
Tested-by: Sascha Wilking
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters