[CLEANUP] Let expected exceptions test for code, not text
When an exception is expected in DependencyUtilityTest, the correct
exception is tested by text. The change of wording of some of those
exceptions caused the related test to fail.

Testing for the exception code instead of the text opens the
possibility to adapt display messages, while the exception code stays
unchanged and stabilizes tests against false positives.

Change-Id: Ib1365c8d7166e522690bb9f8842adc8eb27f635b
Resolves: #60583
Relates: #53572
Releases: 6.3, 6.2
Reviewed-on: http://review.typo3.org/31835
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>