[TASK] Split action buttons into two groups
The buttons in Web>List are split into two different groups: "primary"
and "secondary". Primary buttons are always visible and represent the
default RUD actions (Show, edit, hide, delete).

This patch also modifies the "Extended view" (EV) behaviour, which
affects the secondary buttons:
If EV is disabled, the secondary buttons are invisible, but can be made
visibile by hovering the trigger button. If EV is enabled, the secondary
buttons are always visible.

CSS rules for buttons in tables are now streamlined with the
backend. The blue styling for checked checkboxes in tables is
removed in favor of a cleaner implementation.

Resolves: #63207
Releases: master
Change-Id: Id1bfd42e658e9dbe0a7c13066fed618539fd5ce0
Reviewed-on: http://review.typo3.org/34788
Reviewed-by: Benjamin Kott <info@bk2k.info>
Tested-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>