[!!!][BUGFIX] Add only required classes and class aliases
Additionally we enable the composer class loading by default
if the according autoload.php file is found in Packages/Libraries/

To be able to detect the compatibility extension being installed
which brings class aliases for classes spread over other system extensions,
we need to add functionality that puts the class aliases to typo3conf/
when an extension is activated or deactivated.

This deviates from the approach in 6.2 branch and most likely will vanish
again later once we require composer installation

If the class alias file is present, the aliases will be added to the composer
class loader during TYPO3 bootstrap.

Releases: master, 6.2
Resolves: #64376
Change-Id: I4fd2202abad345c29bcb2b05d3702bdeef9e4474
Reviewed-on: http://review.typo3.org/36644
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>