[BUGFIX] Store dependency errors before further dependency tests
During dependency checking the extensions are checked recursively. If an
depended extension is not yet installed, dependencies for this extension
are tested as well. Unfortunately the second test resets the array with
stored dependency errors which leads to improper dependency handling.
This patch stores the dependency errors before starting further tests
and re-adds them after processing.

Resolves: #66410
Releases: master, 6.2
Change-Id: Ic576c30f6b1f066467c4899ea5e958280c12d37c
Reviewed-on: http://review.typo3.org/38686
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>