[FEATURE] Add possibility to disable TER download in EM
Currently it isn't possible to use the extension manager without any
internet connection.

This patch adds an option in the extension manager settings to switch to
an offline mode. This means, that the navigation in EM is limited to the
list view. Download of extensions and distributions is disabled.
Furthermore fetching the extension list from TER isn't possible anymore.

Resolves: #66597
Releases: master
Change-Id: I68477f2289912765aa8ae558bb7df6e550f4f813
Reviewed-on: http://review.typo3.org/40428
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Oliver Eglseder <oliver.eglseder@in2code.de>
Tested-by: Oliver Eglseder <oliver.eglseder@in2code.de>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>