[TASK] Add notification about composer mode
If the system was installed with composer, the user should better not
use the extension manager. This patch adds a notification to remind the
user that the system might break otherwise.

This is the first step before some extension manager functionality
will be disabled in the following TYPO3 version.

Resolves: #68527
Releases: master
Change-Id: I72a05481d9dda8a72e3962c45b7a6c5d49d27e01
Reviewed-on: http://review.typo3.org/41941
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>