[TASK] Prevent gathering update information on Ter installation
Currently when the extension list from TER is fetched from the database,
all installed extensions are checked for an update. This is because the
TER extension list needs a css class for available extensions but the
wrong method is used to gather the needed information within the
ListController.

Resolves: #70264
Releases: master, 6.2
Change-Id: I99e9bfceb0bf0ed97d0bf26ae927743fd2ca86f4
Reviewed-on: http://review.typo3.org/43692
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>