[BUGFIX] Reenable JS in Extension Manager
Due to commit https://review.typo3.org/40936/ the JavaScript loaded by
the Extension Manager breaks because jQuery is run in noConflict() mode.

This patch adds a new option within the ContainerViewHelper and
PageRendererViewHelper to be able to set an alternative namespace for
jQuery loading, allowing to have jQuery global again.

Resolves: #71618
Releases: master
Change-Id: Idb63927facee0dfb4797c1dda007033915d7256a
Reviewed-on: https://review.typo3.org/44746
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>