[!!!][FEATURE] Introduce Guzzle for making HTTP Requests
The Guzzle library is added as a composer dependency
instead of HttpRequest to request urls from TYPO3.

Guzzle is encapsulated inside a RequestFactory to do
requests and return PSR-7 compliant Response objects
to deal with content. In the future, Guzzle can also be used
to do HTTP requests asynchronously but there is no API
for that yet as there is no current need for that in the
TYPO3 Core.

GeneralUtility::getUrl() now uses Guzzle under the hood,
thus adding headers like the TYPO3 User Agent by default.

A lot of existing TYPO3_CONF_VARS options are now
removed or merged into Guzzle compliant options which
are independant of the implementation (cURL, stream
wrappers, fopen etc).

Resolves: #70056
Releases: master
Change-Id: Ibd14bba944b1590bae1b12c10f26365f20576475
Reviewed-on: https://review.typo3.org/43508
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>