[TASK] Add button to reload static database data of an extension
This patch re-adds the possibility to manually trigger a reload of
static database data an extension might ships. In the EM list view, an
inconspicuous button is added for each extension which has an own
ext_tables_static+adt.sql file. Furthermore the import is changed, so
that a md5 hash of the file content is stored in the database. If the
extension is updated and the content if the file changes, a more visible
button is shown to indicate an update should be considered.

Resolves: #60748
Releases: master, 7.6
Change-Id: I8ccb01d7e1f98a4adfc70a4d1aa80db4e76bbc9d
Reviewed-on: https://review.typo3.org/51310
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Matthes
Tested-by: Christian Matthes
Tested-by: Joerg Kummer <service@enobe.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>