[TASK] Refactor distribution list view with cards
The card component has been extended and is now more flexible to be
more usefull for different usecases, and is now used in the distribution
listing of the extension manager. We have also changed to Documentation
format of the CSS files to match KSS Knyle Style Sheets format.

Releases: master
Resolves: #82000

Change-Id: I61711dd0ee6c38e3c249d9b744a3ee6b2e24273e
Reviewed-on: https://review.typo3.org/53619
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>