[BUGFIX] Allow empty date/time in strict database modes
Empty date/time values are represented by the FormEngine as empty
string (''). Trying to persist this value in database management systems
using strict mode (which is enabled per default in MySQL 5.7),
will cause an SQL error since field types do not match.

Since it's known for the TYPO3 core which field types are used in
particular database tables, it's also possible to enforce this
type to be casted to integer values. This is done by using the
according TCA evaluation function 'int'.

Resolves: #82756
Releases: master, 8.7
Change-Id: I487caae785a3943916bceebc0c484d58e900a841
Reviewed-on: https://review.typo3.org/56062
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>