[BUGFIX] Ensure fresh logger instance in scheduler tasks
The AbstractTask always had methods setScheduler and
unsetScheduler to avoid serialization of the scheduler instance.
These methods are called by the scheduler API before serializing
(unset) and after unserializing (set).

Use these methods in favor of the __sleep and __wakup calls
to avoid issues with third party subclasses that implement these methods
and to make sure tasks that are already serialized with an outdated
logger instance are properly updated on next run without the need
to clean them up manually.

Releases: master, 9.5
Resolves: #86941
Resolves: #87780
Change-Id: I140973fa952c5c41b8eac966ba3b541420ec060b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60306
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>