[BUGFIX] Prevent ReflectionException when scheduler/reports is unavailable
When symfony/dependency-injection crawls the available classes
a ReflectionException [1] occurs when the base class or interface of an
analyzed class is not available.

This currently happens for scheduler tasks and status reports in composer
mode. The dependency to external base classes/interfaces is missing in the
respective composer.json files and therefore EXT:scheduler and EXT:reports
are not guaranteed to be availble.

Exclude the affected status report and scheduler task classes from
autowiring and autoconfiguration in order to fix this issue.

Note: Dependency injection can still be added for these classes,
but as long as the implicit dependencies to EXT:scheduler and EXT:reports
are not fixed, these classes will need to be wired manually. Example:

  TYPO3\CMS\Reports\Task\SystemStatusUpdateTask:
    autoconfigure: false
    autowire: false
    public: true
    arguments:
      - '@cache.core'

[1] Uncaught ReflectionException:
    Class TYPO3\CMS\Scheduler\Task\AbstractTask not found in
    […]/typo3/sysext/reports/Classes/Task/SystemStatusUpdateTask.php:28

Resolves: #88825
Resolves: #88912
Releases: master
Change-Id: I74113aa87336bbde02b8754639eafc4158dc3717
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61342
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>