[BUGFIX] Add missing `remote` column in TCA
In #92457 a new field `remote` was introduced to
`tx_extensionmanager_domain_model_extension`.

As it was not added to the corresponding TCA,
accessing the value through the extbase model
throws an exception when installing an extension
in Extension Manager.

It's therefore now added to the TCA, enabling the
automatic extbase mapping.

Resolves: #92533
Relates: #92457
Releases: master
Change-Id: I685ce04d18f7c623372ac809cc5a339fc64e31c6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66099
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Benni Mack <benni@typo3.org>