[TASK] Create TER remote instances in symfony service factory
The remotes are not lazy loaded, therefore there is no need to
inject the container instance into RemoteRegistry.

We can pass the remote instances directly into the
registerRemote method, that has the slight benefit
of avoiding roundtrips and the symfony service defintions
can be inlined/optimized (the remote instances are
created withing the generated RemoteRegistry factory).

Resolves: #92722
Related: #92457
Releases: master
Change-Id: Ie8e52a5d48d2d1614d8f554f0854408cc11ec17d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66324
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>