[BUGFIX] Do not call "libxml_disable_entity_loader" in PHP 8
Because the method (finally) is deprecated in PHP 8,
all calls in TYPO3 Core are wrapped in if statements
to avoid deprecation warnings.

PHP 8 effectively is more secure by default,
which is a good thing, but we need to consider this
in our code base, which still supports PHP 7
as well.

Resolves: #93204
Releases: master, 10.4
Change-Id: I18d7e76e3de5cf48cd4c3cab0d68dea4e518f674
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67295
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>