[BUGFIX] Fix extension listing
The listing of extensions in the Extension Manager module
now properly shows the inactive extensions again.

Previous CSS changes didn't apply the inactive extensions
again.

In addition, if an extension does not have an Extension Icon
in the Extension Manager list, it created an <img> tag
with a path to the extension folder typo3conf/ext/my_ext/
resulting in an error in the browsers. This is now
replaced with an empty icon.

Resolves: #93311
Releases: master
Change-Id: I723a0d603b36ad799c0c02869829742f19a69b37
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67462
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>