[TASK] Simplify extension manager xml parser
When dealing with TER extensions in the extension manager,
the TER provides an XML file with extension details that
are parsed and loaded into a database table.

When this has been added a long time ago, two XML parsers
have been added together with a factory to select one.

Since PHP extension 'xml' is a required TYPO3 extension
since a lot of core versions, always the first parser
did the job.

The patch removes the legacy parser and the factory,
injects the parser via DI to the consuming class, merges
the abstract parser class into the now single parser,
moves the namespace, and refactors the
class towards PHP strict typing.

This leads to a significantly simplified codebase, much
easier to maintain and adapt.

Change-Id: I5f64370f943964686717dccb4e39cda1ae9511c9
Resolves: #94220
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69307
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>