[TASK] Remove unused variables throughout core
Due to refactoring or very old code some places in the core define
unused variables or variables, which are immediately overridden.
They are mostly unnecessary. Even in places, where they might be
slightly helpful if defined at the beginning of a function, the benefit
doesn't excuse the existence.

Resolves: #94528
Releases: master
Change-Id: I6071bb0bc96472a4b4b7f679bfabbacfb761f472
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69805
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>