[BUGFIX] Do not add 'offlineMode' setting in extbase by default
Since #94057, ['offlineMode' => false] is added to the settings array
in ActionController by default. This has most likely been added by
mistake, since it is not used in extbase but in ext:extensionmanager.

This patch removes the default 'offlineMode' setting in extbase and
ensures, it is always defined in ext:extensionmanager.

Resolves: #95187
Related: #94057
Releases: master
Change-Id: Id43ea5b4ea598f374193e9119aad1397e9e4aef8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71028
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>