[TASK] Use icon identifier in module configurations
In most module configurations icons with a distinct path are
used. This is not necessary, as the module icons are already
available from the TYPO3.Icons package and therefore the
icon identifier can be used. For EXT:core this was already done.

The previously referenced icons are not used anymore and will
be removed by a separated patch for v12 once v11 is branched.

Resolves: #95888
Releases: master
Change-Id: Ic113fc9f40ce80059b65e1290afe70673ba06a33
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72070
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>