[FEATURE] Introduce TCA type "datetime"
In our process of using dedicated TCA types,
the new TCA type "date" is introduced and
replaces "renderType=inputDateTime" of TCA
type "input".

Notable changes to the previous behaviour:

- the "format" option, which was already documented
  but never used in FormEngine, is now mandatory
- TCA option "max" is no longer evaluated
- The only allowed eval's are "null" and "int"

The corresponding FormEngine element class is
renamed from InputDateTimeElement to DatetimeElement.
This is covered in the ClassAliasMap.

Note: The "eval=int" seems to be superfluous since
the field value should always be stored as int in
case no native type ("dbType") is used. However,
this would be breaking and is therefore done in
a separate followup patch. This has already been
kickstarted in https://review.typo3.org/c/Packages/TYPO3.CMS/+/74027/12.

Resolves: #97232
Releases: main
Change-Id: I94c6d12759cd7787e6185ab36d2fcd60f5dc15df
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74027
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>