[TASK] Replace FontAwesome icons with typo3 icons
To provide a consistent look&feel and to eventually phase out
FontAwesome in the TYPO3 backend, all remaining fa-* icons are replaced
with their counterparts from the @typo3/icons package.

The CSS class `icon-emphasized` is introduced as a drop-in replacement
for the `fa-stack` construct:

```
<span class="icon-emphasized">
    <core:icon identifier="my-icon-identifier" size="small"/>
</span>
```

The removal of the FontAwesome integration will follow in a separate
patch.

Resolves: #97877
Releases: main
Change-Id: I0e9ca1e4a918b3f44969ea9886ec0cd1e56d78be
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75041
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>