[BUGFIX] Store search terms in local extension list
With the removal of datatables in #86343 the search terms used to filter
tables were not stored anymore, leading to unfiltered tables after
requesting pages with such again.

This is found to be annoying in the extension manager when many
extensions with the same prefix need to get managed.

The search terms are now stored in the browser session and the local
extension list is filtered again when visiting the extension manager.

Resolves: #93628
Related: #86343
Releases: main, 11.5
Change-Id: Ib99e1fbf0ad420826164f8fed019c5235f8c43d5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75155
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>