[TASK] Show modal before re-import of static data
Extensions can provide static data using the file
"ext_tables_static+adt.sql".
If this file exists, a button is automatically created
in the Extension Manager.
As soon as you click the button, the static SQL data
will be imported. Existing data will be dropped.

A new modal now warns the user before executing the file,
with buttons to start or abort the task.

Resolves: #100118
Releases: main, 12.4
Change-Id: Ie33a89d6957428f390edeaa2700639bc13ac8b7d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80327
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>