[BUGFIX] Redefine database field review_state of ExtensionManager
The field was defined as type='check', which gets autogenerated
since #101653 as a positive int value. The field however contains
values in the negative range, rendering it unfit to be represented
as a checkbox field.
Redefining the field as a numeric input field and restoring the
sql definiton avoids the possible data loss and handling errors.

Resolves: #101895
Related: #101653
Releases: main
Change-Id: I8e2086ee6e364b0ee8594f898988164ffa25f8b5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80942
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Thomas Hohn <tho@gyldendal.dk>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Wand <wand@itaw.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>