[TASK] Avoid implicitly nullable class method parameter in `EXT:extensionmanager`
With PHP 8.4 marking method parameter implicitly nullable
is deprecated and will emit a `E_DEPRECATED` warning. One
recommended way to resolve this, is making it explicitly
nullable using the `?` nullable operator or adding a null
tyype to an union type definition. [1]

This prepares the way towards PHP 8.4 compatibility.

[1] https://php.watch/versions/8.4/implicitly-marking-parameter-type-nullable-deprecated

Resolves: #104247
Releases: main, 12.4, 11.5
Change-Id: I71d8f59eaba6dd22b00f8eb7400f2b49d748e430
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84991
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>